requirements.txt
dotenv
discord.py=>2.5.2
loguru=>0.7.3
python-box=>7.3.2
